feat(filedb): implement cache eviction and metadata management for torrents
- Introduced a new `FileDB.Cron` class to handle periodic cache eviction of outdated entries based on configuration settings.
- Added `Cron` and `CronFast` methods for managing cache entries, improving memory efficiency and performance.
- Created `FileDB.MasterDb`, `FileDB.Metadata`, and `FileDB.UrlParsing` classes to manage torrent metadata, URL parsing, and database interactions.
- Refactored existing methods to streamline torrent information retrieval and update processes, enhancing overall data integrity and access speed.
